Move helper functions to seperate files
